+++ cairo1-run/src/cairo_run.rs
@@ -499,7 +499,9 @@ fn create_entry_code(
                 BuiltinName::keccak
                 | BuiltinName::ecdsa
                 | BuiltinName::output
-                | BuiltinName::range_check96 => return fp_loc,
+                | BuiltinName::range_check96
+                | BuiltinName::add_mod
+                | BuiltinName::mul_mod => return fp_loc,
             };
             signature
                 .ret_types