+++ vm/src/serde/deserialize_program.rs
@@ -19,6 +19,7 @@ use crate::{
 
 use crate::utils::PRIME_STR;
 use crate::vm::runners::builtin_runner::SEGMENT_ARENA_BUILTIN_NAME;
+use crate::vm::runners::builtin_runner::{ADD_MOD_BUILTIN_NAME, MUL_MOD_BUILTIN_NAME};
 use crate::Felt252;
 use crate::{
     serde::deserialize_utils,
@@ -57,6 +58,8 @@ pub enum BuiltinName {
     poseidon,
     segment_arena,
     range_check96,
+    add_mod,
+    mul_mod,
 }
 
 impl BuiltinName {
@@ -72,6 +75,8 @@ impl BuiltinName {
             BuiltinName::poseidon => POSEIDON_BUILTIN_NAME,
             BuiltinName::segment_arena => SEGMENT_ARENA_BUILTIN_NAME,
             BuiltinName::range_check96 => RANGE_CHECK_96_BUILTIN_NAME,
+            BuiltinName::add_mod => ADD_MOD_BUILTIN_NAME,
+            BuiltinName::mul_mod => MUL_MOD_BUILTIN_NAME,
         }
     }
 }