+++ vm/src/types/instance_definitions/builtins_instance_def.rs
@@ -1,9 +1,11 @@
+use super::mod_instance_def::ModInstanceDef;
 use super::{
     bitwise_instance_def::BitwiseInstanceDef, ec_op_instance_def::EcOpInstanceDef,
     ecdsa_instance_def::EcdsaInstanceDef, keccak_instance_def::KeccakInstanceDef,
     pedersen_instance_def::PedersenInstanceDef, poseidon_instance_def::PoseidonInstanceDef,
     range_check_instance_def::RangeCheckInstanceDef,
 };
+
 use serde::Serialize;
 
 #[derive(Serialize, Debug, PartialEq)]
@@ -17,6 +19,8 @@ pub(crate) struct BuiltinsInstanceDef {
     pub(crate) keccak: Option<KeccakInstanceDef>,
     pub(crate) poseidon: Option<PoseidonInstanceDef>,
     pub(crate) range_check96: Option<RangeCheckInstanceDef>,
+    pub(crate) add_mod: Option<ModInstanceDef>,
+    pub(crate) mul_mod: Option<ModInstanceDef>,
 }
 
 impl BuiltinsInstanceDef {
@@ -31,6 +35,8 @@ impl BuiltinsInstanceDef {
             keccak: None,
             poseidon: None,
             range_check96: None,
+            add_mod: None,
+            mul_mod: None,
         }
     }
 
@@ -45,6 +51,8 @@ impl BuiltinsInstanceDef {
             keccak: None,
             poseidon: None,
             range_check96: None,
+            add_mod: None,
+            mul_mod: None,
         }
     }
 
@@ -59,6 +67,8 @@ impl BuiltinsInstanceDef {
             keccak: None,
             poseidon: None,
             range_check96: None,
+            add_mod: None,
+            mul_mod: None,
         }
     }
 
@@ -73,6 +83,8 @@ impl BuiltinsInstanceDef {
             keccak: None,
             poseidon: None,
             range_check96: None,
+            add_mod: None,
+            mul_mod: None,
         }
     }
 
@@ -87,6 +99,8 @@ impl BuiltinsInstanceDef {
             keccak: None,
             poseidon: Some(PoseidonInstanceDef::default()),
             range_check96: None,
+            add_mod: None,
+            mul_mod: None,
         }
     }
 
@@ -101,6 +115,8 @@ impl BuiltinsInstanceDef {
             keccak: Some(KeccakInstanceDef::new(Some(2048), vec![200; 8])),
             poseidon: Some(PoseidonInstanceDef::default()),
             range_check96: None,
+            add_mod: None,
+            mul_mod: None,
         }
     }
 
@@ -115,6 +131,8 @@ impl BuiltinsInstanceDef {
             keccak: None,
             poseidon: Some(PoseidonInstanceDef::new(Some(8))),
             range_check96: None,
+            add_mod: None,
+            mul_mod: None,
         }
     }
 
@@ -129,6 +147,14 @@ impl BuiltinsInstanceDef {
             keccak: Some(KeccakInstanceDef::new(Some(2048), vec![200; 8])),
             poseidon: Some(PoseidonInstanceDef::new(Some(256))),
             range_check96: Some(RangeCheckInstanceDef::new(Some(8))),
+            #[cfg(feature = "mod_builtin")]
+            add_mod: Some(ModInstanceDef::new(Some(128), 1, 96)),
+            #[cfg(feature = "mod_builtin")]
+            mul_mod: Some(ModInstanceDef::new(Some(256), 1, 96)),
+            #[cfg(not(feature = "mod_builtin"))]
+            add_mod: None,
+            #[cfg(not(feature = "mod_builtin"))]
+            mul_mod: None,
         }
     }
 
@@ -143,6 +169,8 @@ impl BuiltinsInstanceDef {
             keccak: None,
             poseidon: None,
             range_check96: None,
+            add_mod: None,
+            mul_mod: None,
         }
     }
 
@@ -157,6 +185,14 @@ impl BuiltinsInstanceDef {
             keccak: None,
             poseidon: None,
             range_check96: None,
+            #[cfg(feature = "mod_builtin")]
+            add_mod: Some(ModInstanceDef::new(None, 1, 96)),
+            #[cfg(feature = "mod_builtin")]
+            mul_mod: Some(ModInstanceDef::new(None, 1, 96)),
+            #[cfg(not(feature = "mod_builtin"))]
+            add_mod: None,
+            #[cfg(not(feature = "mod_builtin"))]
+            mul_mod: None,
         }
     }
 }