+++ vm/src/types/instance_definitions/mod.rs
@@ -5,6 +5,8 @@ pub mod diluted_pool_instance_def;
 pub mod ec_op_instance_def;
 pub mod ecdsa_instance_def;
 pub mod keccak_instance_def;
+#[allow(unused)]
+pub mod mod_instance_def;
 pub mod pedersen_instance_def;
 pub mod poseidon_instance_def;
 pub mod range_check_instance_def;