+++ vm/src/vm/runners/builtin_runner/poseidon.rs
@@ -110,10 +110,6 @@ impl PoseidonBuiltinRunner {
         Ok(self.cache.borrow().get(&address).map(|x| x.into()))
     }
 
-    pub fn get_memory_segment_addresses(&self) -> (usize, Option<usize>) {
-        (self.base, self.stop_ptr)
-    }
-
     pub fn get_used_cells(&self, segments: &MemorySegmentManager) -> Result<usize, MemoryError> {
         segments
             .get_segment_used_size(self.base())
@@ -126,43 +122,6 @@ impl PoseidonBuiltinRunner {
     ) -> Result<usize, MemoryError> {
         let used_cells = self.get_used_cells(segments)?;
         Ok(div_ceil(used_cells, self.cells_per_instance as usize))
-    }
-
-    pub fn final_stack(
-        &mut self,
-        segments: &MemorySegmentManager,
-        pointer: Relocatable,
-    ) -> Result<Relocatable, RunnerError> {
-        if self.included {
-            let stop_pointer_addr = (pointer - 1)
-                .map_err(|_| RunnerError::NoStopPointer(Box::new(POSEIDON_BUILTIN_NAME)))?;
-            let stop_pointer = segments
-                .memory
-                .get_relocatable(stop_pointer_addr)
-                .map_err(|_| RunnerError::NoStopPointer(Box::new(POSEIDON_BUILTIN_NAME)))?;
-            if self.base as isize != stop_pointer.segment_index {
-                return Err(RunnerError::InvalidStopPointerIndex(Box::new((
-                    POSEIDON_BUILTIN_NAME,
-                    stop_pointer,
-                    self.base,
-                ))));
-            }
-            let stop_ptr = stop_pointer.offset;
-            let num_instances = self.get_used_instances(segments)?;
-            let used = num_instances * self.cells_per_instance as usize;
-            if stop_ptr != used {
-                return Err(RunnerError::InvalidStopPointer(Box::new((
-                    POSEIDON_BUILTIN_NAME,
-                    Relocatable::from((self.base as isize, used)),
-                    Relocatable::from((self.base as isize, stop_ptr)),
-                ))));
-            }
-            self.stop_ptr = Some(stop_ptr);
-            Ok(stop_pointer_addr)
-        } else {
-            self.stop_ptr = Some(0);
-            Ok(pointer)
-        }
     }
 
     pub fn air_private_input(&self, memory: &Memory) -> Vec<PrivateInput> {
@@ -231,7 +190,7 @@ mod tests {
     #[test]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn final_stack() {
-        let mut builtin = PoseidonBuiltinRunner::new(Some(10), true);
+        let mut builtin: BuiltinRunner = PoseidonBuiltinRunner::new(Some(10), true).into();
 
         let mut vm = vm!();
 
@@ -255,7 +214,7 @@ mod tests {
     #[test]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn final_stack_error_stop_pointer() {
-        let mut builtin = PoseidonBuiltinRunner::new(Some(10), true);
+        let mut builtin: BuiltinRunner = PoseidonBuiltinRunner::new(Some(10), true).into();
 
         let mut vm = vm!();
 
@@ -283,7 +242,7 @@ mod tests {
     #[test]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn final_stack_error_when_not_included() {
-        let mut builtin = PoseidonBuiltinRunner::new(Some(10), false);
+        let mut builtin: BuiltinRunner = PoseidonBuiltinRunner::new(Some(10), false).into();
 
         let mut vm = vm!();
 
@@ -307,7 +266,7 @@ mod tests {
     #[test]
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn final_stack_error_non_relocatable() {
-        let mut builtin = PoseidonBuiltinRunner::new(Some(10), true);
+        let mut builtin: BuiltinRunner = PoseidonBuiltinRunner::new(Some(10), true).into();
 
         let mut vm = vm!();
 
@@ -443,7 +402,7 @@ mod tests {
     fn get_air_private_input() {
         let builtin: BuiltinRunner = PoseidonBuiltinRunner::new(None, true).into();
 
-        let memory = memory![
+        let segments = segments![
             ((0, 0), 0),
             ((0, 1), 1),
             ((0, 2), 2),
@@ -458,7 +417,7 @@ mod tests {
             ((0, 11), 11)
         ];
         assert_eq!(
-            builtin.air_private_input(&memory),
+            builtin.air_private_input(&segments),
             (vec![
                 PrivateInput::PoseidonState(PrivateInputPoseidonState {
                     index: 0,